Create README.md
Facebook Restored brings back the chat function many know and love; "Turn on/off chat for all contacts except". With this extension, you will once again know the joy of being able to turn your chat off for that one nagging individual, or keeping it on for that special one.